Implemented createAccount() in Bank.cpp and added a new member function to Bank
- createAccount() works in creating either a Savings or Checking Account
- used std::make_unique<T>() and std::unique_ptr<T> for dynamic memory management
- accountExists() added to make sure no duplicate accounts are made
- accountExists takes one parameter which is the account number

+++ include/CheckingAccount.hpp
@@ -4,9 +4,9 @@
 
 class CheckingAccount : public Account{
 private:
-    double overdraftLimit;
+    int overdraftLimit;
 public:
-    CheckingAccount(int accNum, const std::string& name, double balance, double limit) : Account(accNum, name, balance), overdraftLimit(limit) {};
+    CheckingAccount(int accNum, const std::string& name, double balance, int limit) : Account(accNum, name, balance), overdraftLimit(limit) {};
     
     void deposit(double amount) override;
     void withdraw(double amount) override;
